refactor: organizing main menu into a specific service.
